rag/pyproject.toml
[tool.poetry]
name = "rag"
version = "0.1.0"
description = ""
authors = ["Henrik Seidel <heseber+github@mailbox.org>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.3.17"
langgraph = "^0.2.69"
black = "^25.1.0"
ruff = "^0.9.4"
langchain-openai = "^0.3.3"
langchain-anthropic = "^0.3.5"
beautifulsoup4 = "^4.13.1"
langchainhub = "^0.1.21"
langchain-community = "^0.3.16"
tavily-python = "^0.5.0"
python-dotenv = "^1.0.1"
isort = "^6.0.0"
pytest = "^8.3.4"
langchain-chroma = "^0.2.1"
pytest-dotenv = "^0.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
